+++ src/testResultParser.ts
@@ -1,9 +1,10 @@
-import {ETestStatus, ITestCase, ITestClass} from './Interfaces';
 
+import {ETestStatus, ITestCase, ITestClass} from './Interfaces';
+import * as vscode from 'vscode';
 const TerminalInfoRe = /([0-9]+) passed, ([0-9]+) failed, ([0-9]+) ignored, ([0-9]+) warnings/;
 const TerminalClsRe =  /\> ([A-Za-z0-9]+)\.\.\./ ;
 const TerminalItemResultRe = /\| ([A-Za-z0-9_]+)(\([\[\],a-zA-Z1-9\.\b ]+\))*\.\.\. (passed|failed|ignored)./;
-const TerminalFailReasonRe = /\- ([A-Za-z\s'1-9_]+)\./;
+const TerminalFailReasonRe = /\- ([A-Za-z\s'1-9_,]+)\./;
 
 
 export const parseResult = (text: string) => {
@@ -12,7 +13,7 @@ export const parseResult = (text: string) => {
     let currentClassIdx:number=-1;
 
     let passed = 0;
-    let failed = 0;
+    let failed: {count:number,map:Map<ITestCase,string>} = {count: 0, map: new Map()};
     let ignored = 0;
     let warnings = 0;
 
@@ -25,7 +26,7 @@ export const parseResult = (text: string) => {
         if (infoMatch) {
             console.info("info", lineNo);
             passed = parseInt(infoMatch[1], 10);
-            failed = parseInt(infoMatch[2], 10);
+            failed.count = parseInt(infoMatch[2], 10);
             ignored = parseInt(infoMatch[3], 10);
             warnings = parseInt(infoMatch[4], 10);
             continue;
@@ -45,8 +46,7 @@ export const parseResult = (text: string) => {
         if (itemMatch && currentClassIdx !== -1) {
             console.info("item", lineNo);
             let status:ETestStatus = ETestStatus.Unknown;
-
-            switch (itemMatch[2]) {
+            switch (itemMatch[3]) {
                 case "passed":
                     status = ETestStatus.Passed;
                     break;
@@ -58,7 +58,19 @@ export const parseResult = (text: string) => {
                     break;
             }
             const testName = itemMatch[1];
-            const newTestCase: ITestCase = { itemName:testName, status };
+            const args = itemMatch[2];
+            const newTestCase: ITestCase = { itemName:testName, status, args};
+
+            if (status === ETestStatus.Failed) {
+                const reason=lines[lineNo+1];
+                console.log(reason);
+                const ReasonMatch = reason.match(TerminalFailReasonRe);
+                if (!ReasonMatch) {
+                    vscode.window.showErrorMessage("test fail reason not found");
+                    return;
+                }
+                failed.map.set(newTestCase,ReasonMatch[0]);
+            }
             testedClasses[currentClassIdx].tests.push(newTestCase);
             continue;
         }