+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to the "confirma-testadapter" extension will be documented i
 ## [Unreleased]
 
 - Initial release
-- compatible with godot-confirma v0.6.1-beta
+- compatible with godot-confirma v0.7.0-beta
 
 ## [0.0.0 09.07.2024]
 - begining of the project