+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 All notable changes to the "confirma-testadapter" extension will be documented in this file.
 
-
 ## [Unreleased]
+## [0.1.0 beta 15.08.2024]
 
 - Initial release
 - compatible with godot-confirma v0.7.0-beta